+++ src/Mercado.java
@@ -1,20 +1,20 @@
 package src;
 
 abstract class Mercado extends Mercearia{
-    private float areaCorredores;
+    private double areaCorredores;
 
-    public Mercado(String nome, Localizacao local, String distrito, float faturacaoMedia, float custoLimpezaAnual, float areaCorredores) {
+    public Mercado(String nome, Localizacao local, String distrito, double faturacaoMedia, double custoLimpezaAnual, double areaCorredores) {
         super(nome, local, distrito, faturacaoMedia, custoLimpezaAnual);
         this.areaCorredores = areaCorredores;
         this.tipo = "Mercado";
         this.subCategoria = "Mercado";
     }
 
-    public float getAreaCorredores() {
+    public double getAreaCorredores() {
         return areaCorredores;
     }
 
-    public void setAreaCorredores(float areaCorredores) {
+    public void setAreaCorredores(double areaCorredores) {
         this.areaCorredores = areaCorredores;
     }
 
@@ -27,7 +27,7 @@ abstract class Mercado extends Mercearia{
     }
 
     @Override
-    public float receitaAnual() {
+    public double receitaAnual() {
         return areaCorredores*getFaturacaoMedia();}//faturacaoMedia é a faturacao media anual por m^2
 
 }
