+++ src/Pastelaria.java
@@ -1,26 +1,26 @@
 package src;
 
 public class Pastelaria extends Restauracao {
-    private float numMedBolos;
+    private double numMedBolos;
 
-    public Pastelaria(String nome, Localizacao local, String distrito, float faturacaoMedia, int numEmpMesa, float salarioMedAnual, float numMedClientesDiario, float numMedBolos) {
+    public Pastelaria(String nome, Localizacao local, String distrito, double faturacaoMedia, int numEmpMesa, double salarioMedAnual, double numMedClientesDiario, double numMedBolos) {
         super(nome, local, distrito, faturacaoMedia, numEmpMesa, salarioMedAnual, numMedClientesDiario);
         this.numMedBolos = numMedBolos;
         this.tipo = "Pastelaria";
     }
 
-    public float getNumMedBolos() {
+    public double getNumMedBolos() {
         return numMedBolos;
     }
 
-    public void setNumMedBolos(float numMedBolos) {
+    public void setNumMedBolos(double numMedBolos) {
         this.numMedBolos = numMedBolos;
     }
-    public float faturaAnualPBoloPDia(){
+    public double faturaAnualPBoloPDia(){
         return getFaturacaoMedia()/365;
     }
     @Override
-    public float receitaAnual(){
+    public double receitaAnual(){
         return numMedBolos*getFaturacaoMedia();
     }//FaturacaoMedia anual por bolo
 }
