+++ src/Restaurante.java
@@ -4,7 +4,7 @@ public abstract class Restaurante extends Restauracao{
     private int numDiasFuncAnual;
     private int numMesasInteriores;
 
-    public Restaurante(String nome, Localizacao local, String distrito, float faturacaoMedia, int numEmpMesa, float salarioMedAnual, float numMedClientesDiario, int numDiasFuncAnual, int numMesasInteriores) {
+    public Restaurante(String nome, Localizacao local, String distrito, double faturacaoMedia, int numEmpMesa, double salarioMedAnual, double numMedClientesDiario, int numDiasFuncAnual, int numMesasInteriores) {
         super(nome, local, distrito, faturacaoMedia, numEmpMesa, salarioMedAnual, numMedClientesDiario);
         this.tipo = "Restaurante";
         this.subCategoria = "Restaurante";
